move animation utils to main module
